+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 Starting from v2.2.5, all notable changes to this project will be documented in this file.
 
+## v2.7.0
+
+ - Support SeedLink buffer size customization
+ - Basic implementation of SeedLink buffer file
+
 ## v2.6.1
 
  - Fixed a depencency issue in MiniSEED data processing